python/flask/projeto_estudo/requirements.txt
alembic==1.17.2
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.5
packaging==25.0
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.4
WTForms==3.2.1
